+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_commenting_analyzer_ignores/examples/example.dart
@@ -1,0 +1,18 @@
+// ignore_for_file: some_rule
+void main() {
+  // ignore: deprecated_member_use
+  final map = Map(); // LINT
+
+  // ignore: deprecated_member_use, long-method
+  final set = Set(); // LINT
+
+  // Ignored for some reasons
+  // ignore: deprecated_member_use
+  final list = List();
+
+  // ignore: deprecated_member_use same line ignore
+  final queue = Queue();
+
+  // ignore: deprecated_member_use, long-method multiple same line ignore
+  final linkedList = LinkedList();
+}